assets_tutorial: move Daemon into its own file
